+++ modules/admin/src/messages/pt/admin.php
@@ -270,4 +270,9 @@ return [
     'upload_err_message_6' => 'Missing a temporary folder.',
     'upload_err_message_7' => 'Failed to write file to disk.',
     'upload_err_message_8' => 'A PHP extension stopped the file upload.',
+    'model_config_atr_name' => 'Identifier',
+    'model_config_atr_name_hint' => 'Unique identifier for this variable. You can access system config with Config::get(identifier) in your application.',
+    'model_config_atr_value' => 'Value',
+    'model_config_atr_is_system' => 'Is system',
+    'menu_system_item_config' => 'Config variables'
 ];